+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Fixes issue where messages sent from Poodle always appear at end of discussion
 - Changes accent color to light blue
 - Proper activity stream: indicates who did what, and at what time
+- Adds 'Document' object type: can be edited by anyone
 
 ## 2015-08-26